+++ EduSohoAndroidClient/edusoho_v3/src/main/java/com/edusoho/kuozhi/v3/model/bal/push/ServiceProviderModel.java
@@ -1,7 +1,5 @@
 package com.edusoho.kuozhi.v3.model.bal.push;
 
-import com.edusoho.kuozhi.v3.EdusohoApp;
-import com.google.gson.Gson;
 
 /**
  * Created by howzhi on 15/9/24.
@@ -15,7 +13,7 @@ public class ServiceProviderModel {
     public String content;
     public String type;
     public String body;
-    public int createdTime;
+    public long createdTime;
 
     public ServiceProviderModel(){
     }
